api/requirements.txt
azure-functions
numpy
matplotlib
scipy
